requirements.txt
google-adk
fastapi
uvicorn
python-dotenv
psycopg2-binary # PostgreSQL driver
SQLAlchemy
httpx          # For making API calls to Telegram